feat: Add Opus 4.1 model and refactor token analytics
- Add Claude Opus 4.1 (anthropic.claude-opus-4-1-20250805-v1:0) support
- Implement Single Source of Truth for model pricing via IPC
- Remove deprecated modelPricing object from renderer
- Refactor useTokenAnalyticsModal.tsx from 1000+ lines to modular structure:
  - Split into lib/analytics/ for reusable logic
  - Create TokenAnalytics/ components with proper separation
  - Implement error handling and loading states
  - Maintain backward compatibility with existing API
- Update pricing calculation to use async IPC calls
- Add comprehensive tests for Opus 4.1 pricing
