+++ src/renderer/src/lib/pricing/modelPricing.test.ts
@@ -1,13 +1,40 @@
-import { test, expect } from '@jest/globals'
+import { test, expect, beforeEach, jest } from '@jest/globals'
 import { calculateCost, formatCurrency } from './modelPricing'
 
-test('should calculate cost for Claude Sonnet 4', () => {
+interface ModelPricing {
+  input: number
+  output: number
+  cacheRead: number
+  cacheWrite: number
+}
+
+// Mock window.api with proper typing
+const mockGetModelPricing = jest.fn<(modelId: string) => Promise<ModelPricing | null>>()
+
+beforeEach(() => {
+  // Reset mock before each test
+  mockGetModelPricing.mockReset()
+
+  // Setup mock window.api
+  ;(global as any).window = {
+    api: {
+      bedrock: {
+        getModelPricing: mockGetModelPricing
+      }
+    }
+  }
+})
+
+test('should calculate cost for Claude Sonnet 4', async () => {
   const modelId = 'anthropic.claude-sonnet-4-20250514-v1:0'
   const inputTokens = 1000
   const outputTokens = 500
   const cacheReadTokens = 200
   const cacheWriteTokens = 100
 
+  const pricing = { input: 0.003, output: 0.015, cacheRead: 0.0003, cacheWrite: 0.00375 }
+  mockGetModelPricing.mockResolvedValue(pricing)
+
   const expectedCost =
     (inputTokens * 0.003 +
       outputTokens * 0.015 +
@@ -15,7 +42,7 @@ test('should calculate cost for Claude Sonnet 4', () => {
       cacheWriteTokens * 0.00375) /
     1000
 
-  const actualCost = calculateCost(
+  const actualCost = await calculateCost(
     modelId,
     inputTokens,
     outputTokens,
@@ -24,15 +51,19 @@ test('should calculate cost for Claude Sonnet 4', () => {
   )
 
   expect(actualCost).toBeCloseTo(expectedCost, 6)
+  expect(mockGetModelPricing).toHaveBeenCalledWith(modelId)
 })
 
-test('should calculate cost for Claude Opus 4', () => {
+test('should calculate cost for Claude Opus 4', async () => {
   const modelId = 'anthropic.claude-opus-4-20250514-v1:0'
   const inputTokens = 1000
   const outputTokens = 500
   const cacheReadTokens = 200
   const cacheWriteTokens = 100
 
+  const pricing = { input: 0.015, output: 0.075, cacheRead: 0.0015, cacheWrite: 0.01875 }
+  mockGetModelPricing.mockResolvedValue(pricing)
+
   const expectedCost =
     (inputTokens * 0.015 +
       outputTokens * 0.075 +
@@ -40,7 +71,7 @@ test('should calculate cost for Claude Opus 4', () => {
       cacheWriteTokens * 0.01875) /
     1000
 
-  const actualCost = calculateCost(
+  const actualCost = await calculateCost(
     modelId,
     inputTokens,
     outputTokens,
@@ -49,15 +80,19 @@ test('should calculate cost for Claude Opus 4', () => {
   )
 
   expect(actualCost).toBeCloseTo(expectedCost, 6)
+  expect(mockGetModelPricing).toHaveBeenCalledWith(modelId)
 })
 
-test('should calculate cost for Claude Opus 4.1', () => {
+test('should calculate cost for Claude Opus 4.1', async () => {
   const modelId = 'anthropic.claude-opus-4-1-20250805-v1:0'
   const inputTokens = 1000
   const outputTokens = 500
   const cacheReadTokens = 200
   const cacheWriteTokens = 100
 
+  const pricing = { input: 0.015, output: 0.075, cacheRead: 0.0015, cacheWrite: 0.01875 }
+  mockGetModelPricing.mockResolvedValue(pricing)
+
   const expectedCost =
     (inputTokens * 0.015 +
       outputTokens * 0.075 +
@@ -65,7 +100,7 @@ test('should calculate cost for Claude Opus 4.1', () => {
       cacheWriteTokens * 0.01875) /
     1000
 
-  const actualCost = calculateCost(
+  const actualCost = await calculateCost(
     modelId,
     inputTokens,
     outputTokens,
@@ -74,14 +109,18 @@ test('should calculate cost for Claude Opus 4.1', () => {
   )
 
   expect(actualCost).toBeCloseTo(expectedCost, 6)
+  expect(mockGetModelPricing).toHaveBeenCalledWith(modelId)
 })
 
-test('should calculate cost for Claude 3.5 Sonnet', () => {
+test('should calculate cost for Claude 3.5 Sonnet', async () => {
   const modelId = 'anthropic.claude-3-5-sonnet-20241022-v2:0'
   const inputTokens = 1000
   const outputTokens = 500
   const cacheReadTokens = 200
   const cacheWriteTokens = 100
+
+  const pricing = { input: 0.003, output: 0.015, cacheRead: 0.0003, cacheWrite: 0.00375 }
+  mockGetModelPricing.mockResolvedValue(pricing)
 
   const expectedCost =
     (inputTokens * 0.003 +
@@ -90,7 +129,7 @@ test('should calculate cost for Claude 3.5 Sonnet', () => {
       cacheWriteTokens * 0.00375) /
     1000
 
-  const actualCost = calculateCost(
+  const actualCost = await calculateCost(
     modelId,
     inputTokens,
     outputTokens,
@@ -99,38 +138,41 @@ test('should calculate cost for Claude 3.5 Sonnet', () => {
   )
 
   expect(actualCost).toBeCloseTo(expectedCost, 6)
+  expect(mockGetModelPricing).toHaveBeenCalledWith(modelId)
 })
 
-test('should calculate cost for Nova models', () => {
+test('should calculate cost for Nova models', async () => {
   const testCases = [
     {
       modelId: 'amazon.nova-pro-v1:0',
-      expected: { input: 0.0008, output: 0.0032, cacheRead: 0.0002, cacheWrite: 0 }
+      pricing: { input: 0.0008, output: 0.0032, cacheRead: 0.0002, cacheWrite: 0 }
     },
     {
       modelId: 'amazon.nova-lite-v1:0',
-      expected: { input: 0.00006, output: 0.00024, cacheRead: 0.000015, cacheWrite: 0 }
+      pricing: { input: 0.00006, output: 0.00024, cacheRead: 0.000015, cacheWrite: 0 }
     },
     {
       modelId: 'amazon.nova-micro-v1:0',
-      expected: { input: 0.000035, output: 0.00014, cacheRead: 0.00000875, cacheWrite: 0 }
+      pricing: { input: 0.000035, output: 0.00014, cacheRead: 0.00000875, cacheWrite: 0 }
     }
   ]
 
-  testCases.forEach(({ modelId, expected }) => {
+  for (const { modelId, pricing } of testCases) {
+    mockGetModelPricing.mockResolvedValue(pricing)
+
     const inputTokens = 1000
     const outputTokens = 500
     const cacheReadTokens = 200
     const cacheWriteTokens = 100
 
     const expectedCost =
-      (inputTokens * expected.input +
-        outputTokens * expected.output +
-        cacheReadTokens * expected.cacheRead +
-        cacheWriteTokens * expected.cacheWrite) /
+      (inputTokens * pricing.input +
+        outputTokens * pricing.output +
+        cacheReadTokens * pricing.cacheRead +
+        cacheWriteTokens * pricing.cacheWrite) /
       1000
 
-    const actualCost = calculateCost(
+    const actualCost = await calculateCost(
       modelId,
       inputTokens,
       outputTokens,
@@ -139,32 +181,47 @@ test('should calculate cost for Nova models', () => {
     )
 
     expect(actualCost).toBeCloseTo(expectedCost, 8)
-  })
+    expect(mockGetModelPricing).toHaveBeenCalledWith(modelId)
+
+    // Reset for next iteration
+    mockGetModelPricing.mockReset()
+  }
 })
 
-test('should return 0 for unknown model', () => {
+test('should return 0 for unknown model', async () => {
   const modelId = 'unknown-model'
-  const cost = calculateCost(modelId, 1000, 500)
+  mockGetModelPricing.mockResolvedValue(null)
+
+  const cost = await calculateCost(modelId, 1000, 500)
 
   expect(cost).toBe(0)
+  expect(mockGetModelPricing).toHaveBeenCalledWith(modelId)
 })
 
-test('should handle zero tokens', () => {
+test('should handle zero tokens', async () => {
   const modelId = 'anthropic.claude-sonnet-4-20250514-v1:0'
-  const cost = calculateCost(modelId, 0, 0, 0, 0)
+  const pricing = { input: 0.003, output: 0.015, cacheRead: 0.0003, cacheWrite: 0.00375 }
+  mockGetModelPricing.mockResolvedValue(pricing)
+
+  const cost = await calculateCost(modelId, 0, 0, 0, 0)
 
   expect(cost).toBe(0)
+  expect(mockGetModelPricing).toHaveBeenCalledWith(modelId)
 })
 
-test('should calculate cost without cache tokens', () => {
+test('should calculate cost without cache tokens', async () => {
   const modelId = 'anthropic.claude-sonnet-4-20250514-v1:0'
   const inputTokens = 1000
   const outputTokens = 500
 
+  const pricing = { input: 0.003, output: 0.015, cacheRead: 0.0003, cacheWrite: 0.00375 }
+  mockGetModelPricing.mockResolvedValue(pricing)
+
   const expectedCost = (inputTokens * 0.003 + outputTokens * 0.015) / 1000
-  const actualCost = calculateCost(modelId, inputTokens, outputTokens)
+  const actualCost = await calculateCost(modelId, inputTokens, outputTokens)
 
   expect(actualCost).toBeCloseTo(expectedCost, 6)
+  expect(mockGetModelPricing).toHaveBeenCalledWith(modelId)
 })
 
 test('should format currency with default settings', () => {
