+++ src/renderer/src/lib/pricing/modelPricing.ts
@@ -1,7 +1,3 @@
-// Electron rendererプロセスからmainプロセスの関数を直接インポートできないため、
-// IPCを通して価格情報を取得するAPIを使用する必要があります。
-// 現在は簡略化のため、直接的な実装を維持しています。
-
 interface ModelPricing {
   input: number
   output: number
@@ -11,59 +7,10 @@ interface ModelPricing {
 
 /**
  * モデルの価格情報を取得
- * TODO: IPCを通してmainプロセスのgetModelPricing()を呼び出すように変更
+ * IPCを通してmainプロセスのgetModelPricing()を呼び出す
  */
-const getModelPricingLocal = (modelId: string): ModelPricing | null => {
-  // Claude 3.7 Sonnet
-  if (modelId.includes('claude-3-7-sonnet')) {
-    return { input: 0.003, output: 0.015, cacheRead: 0.0003, cacheWrite: 0.00375 }
-  }
-  // Claude 3.5 Sonnet
-  if (modelId.includes('claude-3-5-sonnet')) {
-    return { input: 0.003, output: 0.015, cacheRead: 0.0003, cacheWrite: 0.00375 }
-  }
-  // Claude 3.5 Haiku
-  if (modelId.includes('claude-3-5-haiku')) {
-    return { input: 0.0008, output: 0.004, cacheRead: 0.00008, cacheWrite: 0.001 }
-  }
-  // Claude 3 Haiku
-  if (modelId.includes('claude-3-haiku')) {
-    return { input: 0.0008, output: 0.004, cacheRead: 0.00008, cacheWrite: 0.001 }
-  }
-  // Claude 3 Sonnet
-  if (modelId.includes('claude-3-sonnet')) {
-    return { input: 0.003, output: 0.015, cacheRead: 0.0003, cacheWrite: 0.00375 }
-  }
-  // Claude 3 Opus
-  if (modelId.includes('claude-3-opus')) {
-    return { input: 0.015, output: 0.075, cacheRead: 0.0015, cacheWrite: 0.01875 }
-  }
-  // Claude Opus 4.1
-  if (modelId.includes('claude-opus-4-1')) {
-    return { input: 0.015, output: 0.075, cacheRead: 0.0015, cacheWrite: 0.01875 }
-  }
-  // Claude Opus 4
-  if (modelId.includes('claude-opus-4')) {
-    return { input: 0.015, output: 0.075, cacheRead: 0.0015, cacheWrite: 0.01875 }
-  }
-  // Claude Sonnet 4
-  if (modelId.includes('claude-sonnet-4')) {
-    return { input: 0.003, output: 0.015, cacheRead: 0.0003, cacheWrite: 0.00375 }
-  }
-  // Nova Pro
-  if (modelId.includes('nova-pro')) {
-    return { input: 0.0008, output: 0.0032, cacheRead: 0.0002, cacheWrite: 0 }
-  }
-  // Nova Lite
-  if (modelId.includes('nova-lite')) {
-    return { input: 0.00006, output: 0.00024, cacheRead: 0.000015, cacheWrite: 0 }
-  }
-  // Nova Micro
-  if (modelId.includes('nova-micro')) {
-    return { input: 0.000035, output: 0.00014, cacheRead: 0.00000875, cacheWrite: 0 }
-  }
-
-  return null
+const getModelPricing = async (modelId: string): Promise<ModelPricing | null> => {
+  return window.api.bedrock.getModelPricing(modelId)
 }
 
 /**
@@ -75,14 +22,14 @@ const getModelPricingLocal = (modelId: string): ModelPricing | null => {
  * @param cacheWriteTokens キャッシュ書き込みトークン数
  * @returns 計算されたコスト（ドル）
  */
-export const calculateCost = (
+export const calculateCost = async (
   modelId: string,
   inputTokens: number,
   outputTokens: number,
   cacheReadTokens: number = 0,
   cacheWriteTokens: number = 0
-): number => {
-  const pricing = getModelPricingLocal(modelId)
+): Promise<number> => {
+  const pricing = await getModelPricing(modelId)
   if (!pricing) return 0
 
   // 1000トークンあたりの価格で計算し、結果を1000で割る
@@ -93,29 +40,6 @@ export const calculateCost = (
       cacheWriteTokens * pricing.cacheWrite) /
     1000
   )
-}
-
-/**
- * バックワード互換性のためのmodelPricingオブジェクト
- *
- * @deprecated このオブジェクトは削除予定です。calculateCost()関数を直接使用してください。
- */
-export const modelPricing = {
-  // Claude 3 Sonnet
-  '3-7-sonnet': { input: 0.003, output: 0.015, cacheRead: 0.0003, cacheWrite: 0.00375 },
-  '3-5-sonnet': { input: 0.003, output: 0.015, cacheRead: 0.0003, cacheWrite: 0.00375 },
-  // Claude 3 Haiku
-  '3-5-haiku': { input: 0.0008, output: 0.004, cacheRead: 0.00008, cacheWrite: 0.001 },
-  // Claude 3 Opus
-  '3-5-opus': { input: 0.015, output: 0.075, cacheRead: 0.0015, cacheWrite: 0.01875 },
-  // Claude 4 Models
-  'sonnet-4': { input: 0.003, output: 0.015, cacheRead: 0.0003, cacheWrite: 0.00375 },
-  'opus-4': { input: 0.015, output: 0.075, cacheRead: 0.0015, cacheWrite: 0.01875 },
-  'opus-4-1': { input: 0.015, output: 0.075, cacheRead: 0.0015, cacheWrite: 0.01875 },
-  // Nova
-  'nova-pro': { input: 0.0008, output: 0.0032, cacheRead: 0.0002, cacheWrite: 0 },
-  'nova-lite': { input: 0.00006, output: 0.00024, cacheRead: 0.000015, cacheWrite: 0 },
-  'nova-micro': { input: 0.000035, output: 0.00014, cacheRead: 0.00000875, cacheWrite: 0 }
 }
 
 /**
